move raw input to new file
